fix: link composition from changelog
